requirements.txt
pytest~=7.1.2
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
requests~=2.28.1
PyYAML~=6.0
ddt~=1.6.0